requirements.txt
fastapi[standard]>=0.111.0
sqlalchemy>=2.0.35
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
openai>=1.0.0